# html/api.yaml
openapi: 3.0.0
info:
  title: Alertflex API
  description: Alertflex public API
  version: 0.1

servers:
  - url: http://_admin_host:8080
    description: HTTP access
  - url: https://_admin_host:8181
    description: HTTPS access
  

paths:
  /alertflex-ctrl/rest/posture/alerts/{taskId}:
    get:
      summary: Return alerts counters 
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: Task ID
      responses:
        '200':
          description: Return alerts metrics
          content:
            application/json:
              schema:
                title: Posture alerts
                type: object
                properties:
                  criticalAlerts:
                    type: integer
                    description: Num of critical alerts
                  majorAlerts:
                    type: integer
                    description: Num of major alerts
                  minorAlerts:
                    type: integer
                    description: Num of minor alerts
                  infoAlerts:
                    type: integer
                    description: Num of info alerts
                  postureType:
                    type: string
                    enum:
                      - AppSecret
                      - DockerConfig
                      - K8sConfig
                      - AppVuln
                      - DockerVuln
                      - K8sVuln
                      - AppSbom
                      - DockerSbom
                      - CloudFormation
                      - Terraform
                      - KubeHunter
                      - ZAP
                      - Nmap
                      - Nuclei
                      - Nikto
                      - CloudSploit
                      - Semgrep
                      - SonarQube
                    description: |
                      Posture type translation to scanner type:

                      AppSecret - trivy scan
                      
                      DockerConfig - trivy scan
                      
                      K8sConfig - trivy scan
                      
                      AppVuln - trivy scan
                      
                      DockerVuln - trivy scan
                      
                      K8sVuln - trivy scan
                      
                      AppSbom - trivy scan
                      
                      DockerSbom - trivy scan
                      
                      CloudFormation - trivy scan
                      
                      Terraform - trivy scan
                      
                      KubeHunter - kubehunter scan
                      
                      ZAP - zap scan
                      
                      Nmap - nmap scan
                      
                      Nuclei - nuclei scan
                      
                      Nikto - nikto scan

                      CloudSploit - cloudsploit scan

                      Semgrep - semgrep scan

                      SonarQube - sonarqube scan

                  target:
                    type: string
                    description: Target for CloudSploit, KubeHunter, Nmap, Nuclei, Nikto, Semgrep, Sonarqube, Trivy and ZAP 
                  vrn:
                    type: string
                    description: VRN name
                  host:
                    type: string
                    description: Host name where scanner is
                required:
                  - criticalAlerts
                  - majorAlerts
                  - minorAlerts
                  - infoAlerts
        '204':
          description: No Content
        
  /alertflex-ctrl/rest/posture/json/{taskId}:
    get:
      summary: Return report for target from CloudSploit, KubeHunter, Nmap, Nuclei, Nikto, Semgrep, Sonarqube, Trivy and ZAP 
      parameters:
        - in: path
          name: taskId
          schema:
            type: string
          required: true
          description: Task ID
      responses:
        '200':
          description: Return id of task
          content:
            application/json:
              schema:
                title: JSON object
                type: object
        '204':
          description: No Content

  /alertflex-ctrl/rest/posture:
    post:
      summary: Create new posture task
      requestBody:
        content:
          application/json:
            schema:
              title: Posture task
              type: object
              properties:
                delay:
                  type: integer
                  description: 
                alertCorr:
                    type: string
                    enum:
                      - AllFindings
                      - NonConfirmed
                      - OnlyNew
                    description: |
                      Type of correlation of alerts:

                      AllFindings - create alerts for all findings
                      
                      NonConfirmed - create alerts for non confirmed alerts
                      
                      OnlyNew - create only for new alerts
                postureType:
                    type: string
                    enum:
                      - AppSecret
                      - DockerConfig
                      - K8sConfig
                      - AppVuln
                      - DockerVuln
                      - K8sVuln
                      - AppSbom
                      - DockerSbom
                      - CloudFormation
                      - Terraform
                      - KubeHunter
                      - ZAP
                      - Nmap
                      - Nuclei
                      - Nikto
                      - CloudSploit
                      - Semgrep
                      - SonarQube
                    description: |
                      Posture type translation to scanner type:

                      AppSecret - trivy scan
                      
                      DockerConfig - trivy scan
                      
                      K8sConfig - trivy scan
                      
                      AppVuln - trivy scan
                      
                      DockerVuln - trivy scan
                      
                      K8sVuln - trivy scan
                      
                      AppSbom - trivy scan
                      
                      DockerSbom - trivy scan
                      
                      CloudFormation - trivy scan
                      
                      Terraform - trivy scan
                      
                      KubeHunter - kubehunter scan
                      
                      ZAP - zap scan
                      
                      Nmap - nmap scan
                      
                      Nuclei - nuclei scan
                      
                      Nikto - nikto scan

                      CloudSploit - cloudsploit scan

                      Semgrep - semgrep scan

                      SonarQube - sonarqube scan

                target:
                  type: string
                  description: Target for CloudSploit, KubeHunter, Nmap, Nuclei, Nikto, Semgrep, Sonarqube, Trivy and ZAP 
                host:
                  type: string
                  description: Host name where scanner is
                vrn:
                  type: string
                  description: VRN name
                project:
                  type: string
                  description: Project Id
              required:
                - project
                - vrn
                - host
                - postureType
                - alertCorr
                - delay
      responses:
        '200':
          description: Return id of task
          content:
            application/json:
              schema:
                title: Task Id
                type: object
                properties:
                  taskId:
                    type: string
                    description: UUID of task
                required:
                  - taskId
        '404':
          description: Return error
          content:
            application/json:
              schema:
                title: Error
                type: object
                properties:
                  error:
                    type: string
                    description: error of task
                required:
                  - error
                  
  /alertflex-ctrl/rest/vrn/status/{projectId}:
    get:
      summary: Return status of sensors and scanners grid for Project ID
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          required: true
          description: Project ID
      responses:
        '200':
          description: Return status
          content:
            application/json:
              schema:
                title: JSON object
                type: object
        '204':
          description: No Content
          
  /alertflex-ctrl/rest/vrn/filters/{vrnName}:
    post:
      summary: Send filters to VRN
      parameters:
        - in: path
          name: vrnName
          schema:
            type: string
          required: true
          description: VRN name
      requestBody:
        content:
          application/json:
            schema:
              title: Filters for VRN
              type: object
      responses:
        '200':
          description: No Content
        '404':
          description: Return error
          content:
            application/json:
              schema:
                title: Error
                type: object
                properties:
                  error:
                    type: string
                    description: error of task
                required:
                  - error
      